Change chip id field to integer field from smallinteger

+++ raspberry/weatherserver/temperatures/models.py
@@ -3,6 +3,6 @@ from datetime import datetime
 
 
 class Probe(models.Model):
-    node_id = models.SmallIntegerField()
+    node_id = models.IntegerField()
     temperature = models.FloatField()
     timestamp = models.DateTimeField(default=datetime.now)
